requirements.txt
python-telegram-bot==20.7
httpx~=0.25.2
groq==0.4.2
python-dotenv==1.0.1